pyproject.toml
[tool.poetry]
name = "div-growth-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
polars = "^1.4.0"
jax = "^0.4.31"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
